+++ src/components/Nav/Nav.module.css
@@ -20,3 +20,12 @@
   font-weight: 700;
   color: #4e4c4c;
 }
+
+.Navtest {
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+  font-size: 37.9px;
+  font-weight: 700;
+  color: #4e4c4c;
+}
